+++ src/cobib/__init__.py
@@ -140,11 +140,11 @@ cobib _example_config > ~/.config/cobib/config.py
 """
 
 import subprocess
-from pathlib import Path, PurePath
+from pathlib import Path
 
 __version__ = "3.0.0"
 
-if Path(PurePath(__file__).parent.parent.parent / ".git").exists():
+if (Path(__file__).parent.parent.parent / ".git").exists():
     # if installed from source, append HEAD commit SHA to version info as metadata
     proc = subprocess.Popen(["git", "rev-parse", "HEAD"], stdout=subprocess.PIPE)
     git_revision, _ = proc.communicate()