+++ src/cobib/commands/base_command.py
@@ -9,10 +9,10 @@ import os
 import shlex
 import sys
 from abc import ABC, abstractmethod
-from pathlib import Path
 from typing import IO, TYPE_CHECKING, Any, Dict, List, NoReturn, Optional
 
 from cobib.config import config
+from cobib.utils.rel_path import RelPath
 
 LOGGER = logging.getLogger(__name__)
 
@@ -95,7 +95,7 @@ class Command(ABC):
         if not git_tracked and not force:
             return
 
-        file = Path(config.database.file).expanduser().resolve()
+        file = RelPath(config.database.file).path
         root = file.parent
 
         if not (root / ".git").exists():