+++ tests/utils/__init__.py
@@ -1,0 +1,1 @@
+"""coBib's utility tests."""